Rename neuralgcm.PressureLevelModel to neuralgcm.Model
This is much less verbose.

The original thinking was that we would want to introduce a new class (with a different API, e.g., based on NNX) for future NeuralGCM models, and hence wanted to use a less generic name, something descriptive of how this model was trained (on pressure level data).

But I now think we can commit to supporting approximately this API for inference with trained models, albeit by deprecating certain properties like `.gin_config`. The details of valid data formats and so forth can be customized based on model checkpoints.

Alternative name options to emphasize the "inference only" aspect (I am equally happy with these):
TrainedModel
SavedModel
Inferrer
Forecaster

PiperOrigin-RevId: 630857761