Add SimTime class for keeping track of elapsed simulated time.
SimTime stores elapsed time in two parts, in order to avoid precision loss associated with storing time in a single float32 value:
- days: the whole number of days that have elapsed
- fraction: the fraction of a day that has elapsed

Total time can be obtained by adding days and fraction.

SimTime can be incremented by a floating point number of days using the `increment` method. Compounding errors are avoided by rounding to the nearest day when the fraction is close to 1 or 0.

PiperOrigin-RevId: 655336029